Se crea nuberFormat y amountFormat
