+++ frontend/src/utils/numberFormat.js
@@ -1,0 +1,7 @@
+const numberFormat = (input) => {
+    const sanitizedValue = input.replace(/[^0-9.]/g, '');
+    
+    return sanitizedValue
+};
+
+export default numberFormat;
