+++ frontend/src/views/bank/Withdraw.jsx
@@ -11,6 +11,7 @@ import Select from 'react-select';
 import { useWithdrawMoneyMutation } from '../../slices/operationApiSlice';
 import UserAccounts from '../../utils/userAccounts';
 import ConfirmOperationModal from '../../components/ConfirmOperationModal';
+import amountFormat from '../../utils/amountFormat';
 
 const WithdrawMoney = () => {
   useCheckCookies();
@@ -42,24 +43,6 @@ const WithdrawMoney = () => {
     label: "Seleccione una cuenta",
     currency: '$',
   });
-
-  const amountChange = (e) => {
-    const inputValue = e.target.value;
-    const sanitizedValue = inputValue.replace(/[^0-9.]/g, '');
-    const decimalCount = (sanitizedValue.match(/\./g) || []).length;
-
-    if (decimalCount > 1) {
-      const [integerPart, decimalPart] = sanitizedValue.split('.');
-      const formattedValue = `${integerPart}.${decimalPart.slice(0, 2)}`;
-      setAmount(formattedValue);
-    } else if (decimalCount === 1 && sanitizedValue.split('.')[1].length > 2) {
-      const [integerPart, decimalPart] = sanitizedValue.split('.');
-      const formattedValue = `${integerPart}.${decimalPart.slice(0, 2)}`;
-      setAmount(formattedValue);
-    } else {
-      setAmount(sanitizedValue);
-    }
-  };
 
   const handleCloseModal = () => {
     setShow(false);
@@ -113,7 +96,7 @@ const WithdrawMoney = () => {
                 <Form.Label className='fw-bold mb-0'>Importe</Form.Label>
                 <InputGroup>
                   <InputGroup.Text>{currency}</InputGroup.Text>
-                  <Form.Control type='text' inputMode='decimal' placeholder='Ingrese importe' value={amount} onChange={amountChange} disabled={!accountId} />
+                  <Form.Control type='text' inputMode='decimal' placeholder='Ingrese importe' value={amount} onChange={(e) => setAmount(amountFormat(e.target.value))} disabled={!accountId} />
                 </InputGroup>
               </Form.Group>
